19-玩家组件与状态机(bug: 上斜坡会触发瓦片碰撞，x速度设置为0)

+++ src/engine/core/GameApp.cpp
@@ -33,7 +33,7 @@ void GameApp::run() {
     
     while(isRunning) {
         _m_time->update();
-        double delta_time = _m_time->getDeltaTime();
+        float delta_time = _m_time->getDeltaTime();
         _m_input_manager->update();
 
         handleEvenets();
